Update faucetTemp
This class implements and demonstrates the operation of a theoretical automatic faucet with no manual temperature controls. A random initial temperature is generated within a range, and a resulting final temperature is calculated. A hot initial temperature results in a final temperature cooler that a set "desired" temperature (medium ground), and a cold initial temperature results in a final temperature hotter than the "desired". The final temperature is set within certain bounds. Tests can show "passed" results or the old/new temperatures.